fixed overflow issue on windows

+++ HPP/country.hpp
@@ -13,7 +13,7 @@ using namespace std;
 class Country {
     public:
         // constructors!
-        Country(const string NAME, const long int POPULATION);
+        Country(const string NAME, const long long int POPULATION);
         Country& operator=(const Country& OTHER);
         ~Country();
         
@@ -24,7 +24,7 @@ class Country {
         bool canSpread = true;
         
         // public operations
-        void new_day(long int& totalInfected, list<Country>& untouched, list<Country>& infectedCountry,  int& lastID);
+        void new_day(long long int& totalInfected, list<Country>& untouched, list<Country>& infectedCountry,  int& lastID);
         void spread(list<Country>& untouched, list<Country>& infectedCountry, int& lastID);
         void infect();
 
@@ -39,20 +39,20 @@ class Country {
 
     private:
         // private attributes
-        long int population;
+        long long int population;
         int infected;
         int infectRate;
         int spreadRate;
 
         // private operations
-        void infect(long int& totalInfected);
+        void infect(long long int& totalInfected);
         void update_infectRate(list<Country>& infectedCountry);
         void update_spreadRate();
 
 };
 
 
-Country::Country(const string NAME, const long int POPULATION) {
+Country::Country(const string NAME, const long long int POPULATION) {
     name = NAME;
     population = POPULATION;
     infected = 0;
@@ -81,7 +81,7 @@ bool operator==(const Country& COUNTRY, const Country& OTHER) {
 
 Country::~Country() = default;
 
-void Country::new_day(long int& totalInfected, list<Country>& untouched, list<Country>& infectedCountry, int& lastID) {
+void Country::new_day(long long int& totalInfected, list<Country>& untouched, list<Country>& infectedCountry, int& lastID) {
     update_infectRate(infectedCountry);
     infect(totalInfected);
     if(untouched.size() > 0) {
@@ -150,7 +150,7 @@ void Country::spread(list<Country>& untouched, list<Country>& infected, int& las
     }
 }
 
-void Country::infect(long int& totalInfected) {
+void Country::infect(long long int& totalInfected) {
     for(int i = 0; i < infected; i ++) { // for each infected person
         int ODDS = rand_int(0, 1001);
         if((ODDS <= infectRate) && (infected < population)) {
